refactor: split permission tests into separate file
Split permission integration tests into cmd/dump/dump_permission_integration_test.go
to keep permission-specific tests separate from the main dump integration tests.

Changes:
- Created cmd/dump/dump_permission_integration_test.go with all permission test scenarios
- Removed permission tests from cmd/dump/dump_integration_test.go
- All tests continue to pass and verify the permission error fix for issue #32
- Maintains clear separation between general dump tests and permission-specific tests

This provides better organization and makes it easier to locate and maintain
permission-related test cases.
